tests/requirements.txt
pytest==8.1.1
pytest-sugar==1.0.0
selenium==4.4.0
webdriver-manager==4.0.2
python-dotenv==0.20.0
allure-pytest==2.13.5
pytest-screenshot-on-failure==1.0.0